Final changed - Completed

+++ src 2/models/order.ts
@@ -25,7 +25,7 @@ OrderInstance.init(
       defaultValue: [],
     },
     total: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.INTEGER
       },
 
   },
